requirements.txt
streamlit
langchain
langchain-google-genai
PyPDF2
python-dotenv
nest_asyncio